create an asc_desc module
