+++ milli/src/criterion.rs
@@ -3,7 +3,8 @@ use std::str::FromStr;
 
 use serde::{Deserialize, Serialize};
 
-use crate::error::{is_reserved_keyword, Error, UserError};
+use crate::error::{Error, UserError};
+use crate::{AscDesc, Member};
 
 #[derive(Debug, Serialize, Deserialize, Clone, PartialEq, Eq)]
 pub enum Criterion {
@@ -87,103 +88,6 @@ impl FromStr for Criterion {
     }
 }
 
-#[derive(Debug, Serialize, Deserialize, Clone, PartialEq)]
-pub enum Member {
-    Field(String),
-    Geo([f64; 2]),
-}
-
-impl FromStr for Member {
-    type Err = UserError;
-
-    fn from_str(text: &str) -> Result<Member, Self::Err> {
-        match text.strip_prefix("_geoPoint(").and_then(|text| text.strip_suffix(")")) {
-            Some(point) => {
-                let (lat, long) = point
-                    .split_once(',')
-                    .ok_or_else(|| UserError::InvalidReservedAscDescSyntax {
-                        name: text.to_string(),
-                    })
-                    .and_then(|(lat, long)| {
-                        lat.trim()
-                            .parse()
-                            .and_then(|lat| long.trim().parse().map(|long| (lat, long)))
-                            .map_err(|_| UserError::InvalidReservedAscDescSyntax {
-                                name: text.to_string(),
-                            })
-                    })?;
-                Ok(Member::Geo([lat, long]))
-            }
-            None => {
-                if is_reserved_keyword(text) || text.starts_with("_geoRadius(") {
-                    return Err(UserError::InvalidReservedAscDescSyntax {
-                        name: text.to_string(),
-                    })?;
-                }
-                Ok(Member::Field(text.to_string()))
-            }
-        }
-    }
-}
-
-impl fmt::Display for Member {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            Member::Field(name) => f.write_str(name),
-            Member::Geo([lat, lng]) => write!(f, "_geoPoint({}, {})", lat, lng),
-        }
-    }
-}
-
-impl Member {
-    pub fn field(&self) -> Option<&str> {
-        match self {
-            Member::Field(field) => Some(field),
-            Member::Geo(_) => None,
-        }
-    }
-
-    pub fn geo_point(&self) -> Option<&[f64; 2]> {
-        match self {
-            Member::Geo(point) => Some(point),
-            Member::Field(_) => None,
-        }
-    }
-}
-
-#[derive(Debug, Serialize, Deserialize, Clone, PartialEq)]
-pub enum AscDesc {
-    Asc(Member),
-    Desc(Member),
-}
-
-impl AscDesc {
-    pub fn member(&self) -> &Member {
-        match self {
-            AscDesc::Asc(member) => member,
-            AscDesc::Desc(member) => member,
-        }
-    }
-
-    pub fn field(&self) -> Option<&str> {
-        self.member().field()
-    }
-}
-
-impl FromStr for AscDesc {
-    type Err = UserError;
-
-    /// Since we don't know if this was deserialized for a criterion or a sort we just return a
-    /// string and let the caller create his own error
-    fn from_str(text: &str) -> Result<AscDesc, Self::Err> {
-        match text.rsplit_once(':') {
-            Some((left, "asc")) => Ok(AscDesc::Asc(left.parse()?)),
-            Some((left, "desc")) => Ok(AscDesc::Desc(left.parse()?)),
-            _ => Err(UserError::InvalidAscDescSyntax { name: text.to_string() }),
-        }
-    }
-}
-
 pub fn default_criteria() -> Vec<Criterion> {
     vec![
         Criterion::Words,
@@ -215,95 +119,9 @@ impl fmt::Display for Criterion {
 #[cfg(test)]
 mod tests {
     use big_s::S;
-    use AscDesc::*;
-    use Member::*;
     use UserError::*;
 
     use super::*;
-
-    #[test]
-    fn parse_asc_desc() {
-        let valid_req = [
-            ("truc:asc", Asc(Field(S("truc")))),
-            ("bidule:desc", Desc(Field(S("bidule")))),
-            ("a-b:desc", Desc(Field(S("a-b")))),
-            ("a:b:desc", Desc(Field(S("a:b")))),
-            ("a12:asc", Asc(Field(S("a12")))),
-            ("42:asc", Asc(Field(S("42")))),
-            ("_geoPoint(42, 59):asc", Asc(Geo([42., 59.]))),
-            ("_geoPoint(42.459, 59):desc", Desc(Geo([42.459, 59.]))),
-            ("_geoPoint(42, 59.895):desc", Desc(Geo([42., 59.895]))),
-            ("_geoPoint(42, 59.895):desc", Desc(Geo([42., 59.895]))),
-            ("_geoPoint(42.0002, 59.895):desc", Desc(Geo([42.0002, 59.895]))),
-            ("_geoPoint(42., 59.):desc", Desc(Geo([42., 59.]))),
-            ("truc(12, 13):desc", Desc(Field(S("truc(12, 13)")))),
-        ];
-
-        for (req, expected) in valid_req {
-            let res = req.parse::<AscDesc>();
-            assert!(
-                res.is_ok(),
-                "Failed to parse `{}`, was expecting `{:?}` but instead got `{:?}`",
-                req,
-                expected,
-                res
-            );
-            assert_eq!(res.unwrap(), expected);
-        }
-
-        let invalid_req = [
-            ("truc:machin", InvalidAscDescSyntax { name: S("truc:machin") }),
-            ("truc:deesc", InvalidAscDescSyntax { name: S("truc:deesc") }),
-            ("truc:asc:deesc", InvalidAscDescSyntax { name: S("truc:asc:deesc") }),
-            ("42desc", InvalidAscDescSyntax { name: S("42desc") }),
-            ("_geoPoint:asc", InvalidReservedAscDescSyntax { name: S("_geoPoint") }),
-            ("_geoDistance:asc", InvalidReservedAscDescSyntax { name: S("_geoDistance") }),
-            (
-                "_geoPoint(42.12 , 59.598)",
-                InvalidAscDescSyntax { name: S("_geoPoint(42.12 , 59.598)") },
-            ),
-            (
-                "_geoPoint(42.12 , 59.598):deesc",
-                InvalidAscDescSyntax { name: S("_geoPoint(42.12 , 59.598):deesc") },
-            ),
-            (
-                "_geoPoint(42.12 , 59.598):machin",
-                InvalidAscDescSyntax { name: S("_geoPoint(42.12 , 59.598):machin") },
-            ),
-            (
-                "_geoPoint(42.12 , 59.598):asc:aasc",
-                InvalidAscDescSyntax { name: S("_geoPoint(42.12 , 59.598):asc:aasc") },
-            ),
-            (
-                "_geoPoint(42,12 , 59,598):desc",
-                InvalidReservedAscDescSyntax { name: S("_geoPoint(42,12 , 59,598)") },
-            ),
-            (
-                "_geoPoint(35, 85, 75):asc",
-                InvalidReservedAscDescSyntax { name: S("_geoPoint(35, 85, 75)") },
-            ),
-            ("_geoPoint(18):asc", InvalidReservedAscDescSyntax { name: S("_geoPoint(18)") }),
-        ];
-
-        for (req, expected_error) in invalid_req {
-            let res = req.parse::<AscDesc>();
-            assert!(
-                res.is_err(),
-                "Should no be able to parse `{}`, was expecting an error but instead got: `{:?}`",
-                req,
-                res,
-            );
-            let res = res.unwrap_err();
-            assert_eq!(
-                res.to_string(),
-                expected_error.to_string(),
-                "Bad error for input {}: got `{:?}` instead of `{:?}`",
-                req,
-                res,
-                expected_error
-            );
-        }
-    }
 
     #[test]
     fn parse_criterion() {
